Removed deprecated files.
